Initialize with values, filter, return self
We are allowing constructor to have initial values for the queries via variable
length unnamed arguments.

Added a basic filter_by_func method, which will require a user defined
filter_func which accepts a query containing a single element from the results,
and returns a bool.

All methods now return self, the instance of the class. This allows for basic
chaining of the query methods. Since we never return any values besides this,
this approach is somewhat robust.
Methods "fetch" and "fetch_with_attributes" return sets, and do NOT allow
further chaining.